Update wallet state management.
This change introduces a new handling for wallet state that covers the cases for connecting and changing chains. This includes making use of a reducer in 'wallet-state.js' to set values for 'connecting' and 'setting-chain' states that are referenced from other components.

Additionally, the file names are updated to use kebab-case convention for file naming and improve on overall code readability for name references.

This provides a more descriptive state management for wallet connections and the ease of tracking the connections. Also, this promotes code consistency with file naming convention for the whole codebase.