Implement automatic adding of fork to wallet
In the web3OnBoard package, we now have the useEffect hook to automatically add the fork to the wallet when one exists. Also, the `extractUuidFromUrl` function was added in the `injected-wallet-interactions.ts` file to extract UUID from the URL, which is then used to create a more descriptive chain label. The motive behind these changes is to enhance the user experience by eliminating the need to manually add forks to the wallet.