Add fallback to disabled network providers
Implemented a fallback check if the requested network provider is in the disabled list. The goal is to return a provider from the disabled list if it exists. This is to enhance our support for multiple network ids and improve the robustness of the getRpcProvider function, even when the requested network id is currently disabled in the app.