Add connection-cancelled handling and correct createConnector behaviour (#2646)
* Add connection-cancelled handling and correct createConnector behaviour

Added a new 'connection-cancelled' state and modified the createConnector function to handle cases where the wallet is already in the process of connecting or is already connected. This is to avoid unnecessary repeated attempts to connect the wallet, and to handle scenarios where a connection attempt is cancelled. In addition, the return type of createConnector was changed from Promise<void> to Promise<Boolean>, with the result indicating if the connection is successful.

* Improve wallet connection process in useBridgeConnector

* Update web3 onboard connector to handle errors

Modified the method to create a connector which now implements error handling when creating a new connector fails. This is done by catching the error and logging it, while also triggering the 'connection-cancelled' dispatch. This ensures better error tracking and state control when connector creation fails.