Refactor wallet management state with Enums (#2654)
* Refactor wallet management state with Enums

Replaced hardcoded status types and event types in wallet management state with more maintainable, readable and well-organized Enums WalletManagementStateStatus and WalletStateEventType respectively. This makes the code easier to navigate, and less prone to typos or inconsistencies. This refactor should not affect functionality, but improves code readability and maintainability.

* Refactor wallet management state with Enums

Replaced hardcoded status types and event types in wallet management state with more maintainable, readable and well-organized Enums WalletManagementStateStatus and WalletStateEventType respectively. This makes the code easier to navigate, and less prone to typos or inconsistencies. This refactor should not affect functionality, but improves code readability and maintainability.