Improve error logging in jsonRpcBatchProvider
This commit enhances the logging of the batch request errors by including the connection URL. The change will print out not only the request that caused the error and the error itself, but also the URL of the connection it was made to. This extra information should help with debugging, especially when troubleshooting problems in environments with multiple instances of the application.