Remove custom network parameter
This commit removes custom network parameter from the application in files along with their respective tests. The network parameter was removed as we now retrieve this information directly from Web3 rather than depending on application's state/configuration. This improvement makes the way we handle networks in the application more understandable and less error-prone.