Refactored web3-on-board-connector-provider for clarity
In this commit, I have removed some unnecessary code to make the web3OnBoard more readable and maintainable. The 'getNetworksFromPageNetwork' function was moved to a separate file for better organization. The condition to change network was also simplified; it now only checks if the desiredNetworkHexId is different from state.walletNetworkHexId. Some unused functions and excess code were removed from the file to declutter the environment. As a result, the code inside 'connect' method is now cleaner and easier to understand.