Update dependencies for ledgerModule and @walletconnect (#2672)
* Update dependencies for ledgerModule and @walletconnect

This commit updates the Ledger Module to include a projectId parameter. Several dependencies are updated as well, including coinbase/wallet-sdk, ledgerhq/connect-kit-loader, and walletconnect. This is done to keep the project up-to-date with the latest functionality and security patches offered by these dependencies.

* Set WalletConnect version in ledgerModule

Introduced "walletConnectVersion: 2" within ledgerModule for better compatibility and update. By setting a specific version for WalletConnect, we ensure compatibility with future updates of WalletConnect. This will provide a more stable and reliable connection

* Update node engine version in package.json

Changed the required node engine version range to "> = 16.x <= 18.x" in package.json. This change provides wider compatibility for the project with various versions of node, allowing developers with different node versions to contribute to the project without issues. It also ensures stability and compatibility of the project with