Refactor codes related to network handling for better flexibility
This commit comprises a series of changes to improve the handling of different networks across the platform. Previously, many parts of the code were hardcoded with specific network ids. They were replaced by a more dynamic approach, providing greater flexibility in setting and changing the network.

The major alteration is in `use-network-connector.tsx` - a new hook `useLegacyDefaultChain` is implemented for remembering the last utilized chain to ensure consistency across different sessions. Also, the `defaultChainId` attribute is now set to `defaultChain` to fetch the stored value from the localStorage.

Simultaneously, removals of redundant or unused code are performed across several files, and a few files for better organizing codes are added. For instance, `use-safafty-reload.tsx` is created to store the page reload logic.

The structure for function getNetworkRpcEndpoint is also simplified to only accept the network ID argument as there was no need anymore for the argument `isTestnet`. This commit is aimed at simplifying the codebase and reducing the redundancy for easier traceability of codes. In addition, it should make the app more resistant to bugs related to network id changes.