[Network Switcher] Goerli Support (#2677)
* Add support for specifying test networks in wallet connection

This commit adds functionality for including test networks when establishing the wallet connection. An optional parameter 'includeTestNet' has been introduced to multiple components and helper functions concerning wallet connections and network configurations.

The new parameter's state is also included in useEffect dependencies lists, ensuring the correct behavior concerning the component's lifecycle when changing the network status. This extension aids in testing and development activities by enabling the interaction with test networks directly from the user interface.

Furthermore, the reducer logic for handling network change events has been extended. Network changes are now handled more gently: instead of a brute force approach, a method was introduced to decide whether sending a network chain change event is necessary or not.

Console.log() statements were introduced for debugging purposes to help & monitor the correctness of the established context.

Finally, extra precautions were included to handle unwanted state changes or transitions between mainnet and the Goerli test network by implementing a safety reload mechanism. This ensures the application's stability & predictability in terms of state transition between differing network environments.

* Improve error logging in jsonRpcBatchProvider

This commit enhances the logging of the batch request errors by including the connection URL. The change will print out not only the request that caused the error and the error itself, but also the URL of the connection it was made to. This extra information should help with debugging, especially when troubleshooting problems in environments with multiple instances of the application.

* Remove custom network parameter

This commit removes custom network parameter from the application in files along with their respective tests. The network parameter was removed as we now retrieve this information directly from Web3 rather than depending on application's state/configuration. This improvement makes the way we handle networks in the application more understandable and less error-prone.

* Add testNet connection capability

The connecting functionality has been updated to provide optional support for testNet. The Context and functions 'connect' and 'shouldSendChangeNetworkOnConnected' have been modified to accept another parameter 'couldBeConnectedToTestNet'. This change allows users to connect to testNet networks by passing 'includeTestNet={true}' in 'WithWalletConnection'. This added flexibility aids in development and testing scenarios without affecting the production-ready, main network functionality.

* Remove unused 'connect' function to switch networks

* Refactored web3-on-board-connector-provider for clarity

In this commit, I have removed some unnecessary code to make the web3OnBoard more readable and maintainable. The 'getNetworksFromPageNetwork' function was moved to a separate file for better organization. The condition to change network was also simplified; it now only checks if the desiredNetworkHexId is different from state.walletNetworkHexId. Some unused functions and excess code were removed from the file to declutter the environment. As a result, the code inside 'connect' method is now cleaner and easier to understand.

* Added function to toggle between Mainnet and Testnet

A new function "toggleBetweenMainnetAndTestnet" has been added to switch connectivity between Mainnet and Testnet. Changes have been made to web3-on-board-connector-provider.tsx, connected-wallet-state-reducer.ts, useConnection.tsx, NavigationNetworkSwitcher.tsx and wallet-state-event.ts files.

This update allows users to easily swap between the two networks, hence improving the user interaction with our blockchain functionalities. Previously, swapping networks required a manual change of network configurations. This all-in-one toggle simplifies this process increasing efficiency and usability.

* Refactor codes related to network handling for better flexibility

This commit comprises a series of changes to improve the handling of different networks across the platform. Previously, many parts of the code were hardcoded with specific network ids. They were replaced by a more dynamic approach, providing greater flexibility in setting and changing the network.

The major alteration is in `use-network-connector.tsx` - a new hook `useLegacyDefaultChain` is implemented for remembering the last utilized chain to ensure consistency across different sessions. Also, the `defaultChainId` attribute is now set to `defaultChain` to fetch the stored value from the localStorage.

Simultaneously, removals of redundant or unused code are performed across several files, and a few files for better organizing codes are added. For instance, `use-safafty-reload.tsx` is created to store the page reload logic.

The structure for function getNetworkRpcEndpoint is also simplified to only accept the network ID argument as there was no need anymore for the argument `isTestnet`. This commit is aimed at simplifying the codebase and reducing the redundancy for easier traceability of codes. In addition, it should make the app more resistant to bugs related to network id changes.

* Update wallet state to improve network flexibility

The wallet-state-guards.ts file has been updated to better handle changes in network id's. Instead of checking for a 'desiredNetworkHexId' in 'state.pageNetworkHexIds', the change ensures that the current wallet's network id is not included. This enhances the application's flexibility when handling various network changes and ensures consistency across different network ids, thereby reducing potential bugs.

* Fix typo in 'useSafetyReload' function and file name

The 'useSafaftyReload' function and its corresponding file are renamed to 'useSafetyReload' to correct the spelling mistake. This change ensures that the function name reflects its purpose more clearly, improving code readability and maintainability.