Update chainId in AutomationContextProvider
Improved chainId handling in AutomationContextProvider by introducing a check for supported networks. Previously, it was directly assigning context.chainId. Now, it will only allow the assignment of chainId if it is either from the MAINNET or GOERLI network, if not, defaults to MAINNET. This ensures the application only communicates with supported networks.