:bug: Fix token addresses for Optimism (#2692)
* Update tokens addresses selection logic and make getLogs call asynchronous

* Update chainId in AutomationContextProvider

Improved chainId handling in AutomationContextProvider by introducing a check for supported networks. Previously, it was directly assigning context.chainId. Now, it will only allow the assignment of chainId if it is either from the MAINNET or GOERLI network, if not, defaults to MAINNET. This ensures the application only communicates with supported networks.

* Add address validation for Web3 bridge connector

Applied the 'isAddress' validation from the ethers library in the bridge-connector.ts file to verify if the provided account address in 'this.wallet.accounts[0].address' is a valid Ethereum address or not. If it's valid, the 'getAddress' function will normalize the address, if not, it will return an empty string. This change provides stronger error handling and robustness to our Web3 implementation.

* Include Goerli network in tokens contract selection

Modified the condition in tokens contract selection to include the Goerli network, besides the Mainnet network. This change allows our application to interact with contracts on both Mainnet and Goerli test network, thus enhancing testing and development workflow.