Implement network check in AjnaProductController and WalletManager
Introduced a network check in AjnaProductController.tsx and web3-on-board-connector-provider.tsx. Modified the WithConnection component in AjnaProductController to specify MAINNET explicitly and to include TestNet. In web3OnBoard, implemented a useEffect hook to handle change of network if the network of connected wallet does not match the required networks. These changes provide the ability for the Ajna page to switch to the required network when user's wallet is connected to a different network. A proper network check and possible switch increases the seamless user experience and mitigates any contract interaction errors caused due to the wrong network.