Add support for networkId in all Ajna-related functions.
This commit enhances various functions related to Ajna, to accept and use the 'networkId' parameter, to ensure correct functioning across multiple blockchain networks. It updates the functions 'getAjnaEarnData', 'getAjnaPoolsData', 'getAjnaPoolAddress', 'getAjnaPoolData', 'searchAjnaPool' and others, to accept and process the networkId. The changes also involve updating the function calls in various places to provide the required 'networkId'.