"Add historical events tracking in Aave state machine"
This commit adds the ability to track historical events in the Aave state machine. This is achieved by adding a new entry 'PROXY_RECEIVED' and 'HISTORY_UPDATED' into the state machine and connecting it to a new service 'getAaveHistoryEvents' that promises a list of AaveHistoryEvent[]. These changes provide the functionality to track and handle historical events throughout the lifetime of a user's interaction with the Aave protocol, improving transparency and traceability for the end-user. This was necessary for enabling detailed analysis and easier debugging of the Aave protocol interactions.