"Improve AavePosition handling and support for AaveV3 on Vault API (#2748)
The changes made were to improve the handling of AavePositions and to include support for AaveV3 on the Vault API. The `PositionCreated` type was updated to include the `PositionType` which will categorize the position as either 'Borrow', 'Multiply', or 'Earn'. Additionally, changes include alterations to the `create_positions$` function to handle both `aaveV2MainnetPositions$` and `aaveV3MainnetPositions$`. Introduction of the `getApiVaults` method to fetch a list of vaults given certain parameters. Vault API responding url was also updated to accommodate chainId information in request. These changes ensure better segregation of positions and broadens the functionality of the API."