Refactor 'Generate' to 'Borrow' action in DebtInput component
The 'Generate' action in the DebtInput component has been refactored to 'Borrow'. This modification better aligns with the borrowing schema of Aave, and the change is made for more precise user interface. In addition, it's relevant to newly incorporated FAQs related to 'Borrow with Aave' feature to provide a more consistent experience for users.