chore: [sc-11214] & [sc-11268] Updating 'Position Info' tab (#2755)
* "Add 'Borrow with Aave' FAQ content and change reference in Aave strategies

The updated FAQs regarding 'Borrow with Aave' feature have been added, including a comprehensive guide and risk explanation for users. The updates were necessary to provide more information and transparency for users about borrowing process and associated risks.

Moreover, in 'ethereum-aave-v3-strategies.ts', the associated FAQ for positionInfo is updated from 'AaveMultiplyFaq' to 'AaveBorrowFaq' to reflect the new content."

* Refactor 'Generate' to 'Borrow' action in DebtInput component

The 'Generate' action in the DebtInput component has been refactored to 'Borrow'. This modification better aligns with the borrowing schema of Aave, and the change is made for more precise user interface. In addition, it's relevant to newly incorporated FAQs related to 'Borrow with Aave' feature to provide a more consistent experience for users.