feat: [sc-11385] Frontend app is broken when URL contains `www` (#2789)
* Several API calls were simplified by removing unnecessary getConfig imports and using relative paths to fetch data from the backend. This change simplifies the codebase, reduces redundancy, and makes the API calls consistent across different services such as the product hub, follow API, wallet risk API, and more. The basePath variable was also removed from several places as it is now unnecessary. This cleanup was essential to keep the codebase clean, readable, and maintainable.

* Update staticFilesRuntimeUrl to return relative paths

* The commit has updated the static runtime URLs to return relative paths instead of absolute paths. This allows for greater flexibility and lets the application be deployed on different servers without changes to the URLs.

Detailed explanation:
The URLs for the gas price API calls have been changed in this commit to `/api/gasPrice` and `/api/tokensPrices`. This removes the `window.location.origin` beforehand which returned the absolute path. Making these URLs relative allows for more fluid deployment of the website as the server origin does not need to be hard-coded into the URL. The `getTicker` function in the `tokensPrices.tsx` file has also been exported to be used in the `product-hub/index.ts` file. This way, token prices are fetched using this function which allows for better error handling. If