Remove unused token metadata files
The token metadata files for aave-v2.ts, aave-v3.ts and ajna.ts were removed, along with the related references to these files in the code base. These files were not required in this version of the app, as these tokens are no longer supported or the support was moved to a different part of application. This cleanup is part of ongoing effort to keep the code base relevant and easy to maintain.