Add support for  sDAI Aave v3 pairs (#2806)
* Remove unused token metadata files

The token metadata files for aave-v2.ts, aave-v3.ts and ajna.ts were removed, along with the related references to these files in the code base. These files were not required in this version of the app, as these tokens are no longer supported or the support was moved to a different part of application. This cleanup is part of ongoing effort to keep the code base relevant and easy to maintain.

* Add SDAI product pairs and update dependencies

Removed the aave-v3 token metadata as we are refactoring the way tokens metadata is handled. Additionally, added several new product pairs in aaveV3Products.ts involving the 'SDAI' token. This is in order to support new long and borrow strategies for these SDAI pairs. Also, new token addresses have been added to 'get-token-addresses.ts' corresponding to the newly supported tokens.

The '@oasisdex/addresses' and '@oasisdex/dma-library' dependency versions have been updated in 'package.json'. This ensures the installed libraries meet the specific version requirements for the functionalities added.

The AaveV3 strategies have also been updated in 'ethereum-aave-v3-strategies.ts' to integrate the operation of the new SDAI product pairs.

* Fix test & typecheck

* Updated '@oasisdex/addresses' and '@oasisdex/dma-library' to stable versions

This commit updates the versions of '@oasisdex/addresses' and '@oasisdex/dma-library' from release candidate versions ('0.0.49-rc.1' and '0.4.14-rc.1' respectively) to their stable versions ('0.0.49' and '0.4.14'). This was done to ensure stability and reduce potential issues caused by testing versions in the production code.