Added RETH and USDT tokens to blockchain contracts
In this update, RETH (Rocket Pool's Ethereum staking token) and USDT (Tether) have been added to the list of contracts in the blockchain service. This was done to extend the range of tokens that the platform can interact with.

Also, the functions used to assert the existence of tokens for the AaveV3 protocol have been modified to accept specific tokens, providing a more granular control over the contracts.

Further, to raise more meaningful error messages during contract assertions, a list of not found token definitions is now provided. This will aid during debugging processes when certain tokens are missing from the active chain.