feat: [sc-11622] [Staging] regression bugs for sdai release (#2813)
* Added RETH and USDT tokens to blockchain contracts

In this update, RETH (Rocket Pool's Ethereum staking token) and USDT (Tether) have been added to the list of contracts in the blockchain service. This was done to extend the range of tokens that the platform can interact with.

Also, the functions used to assert the existence of tokens for the AaveV3 protocol have been modified to accept specific tokens, providing a more granular control over the contracts.

Further, to raise more meaningful error messages during contract assertions, a list of not found token definitions is now provided. This will aid during debugging processes when certain tokens are missing from the active chain.

* Refactor liquidity providers settings for Optimism and Ethereum

This modifies the handling of default exchange protocols for Optimism and Ethereum. Instead of having a single list of default protocols (defaultExchangeProtocols), separate constant arrays for Ethereum (ETHEREUM_MAINNET_DEFAULT_PROTOCOLS) and Optimism (OPTIMIMS_DEFAULT_PROCOTOLS) are now used.

A dynamic pattern matching structure is also introduced in the 'getOneInchCall' method to choose the appropriate protocol list according to the network being used. Several hardcoded Optimism liquidity providers were removed and added to the new 'OPTIMIMS_DEFAULT_PROCOTOLS' array. This comes as a response to the need for more flexibility and clearer coding standards within the project.

Additionally, some protocols from 'ETHEREUM_MAINNET_DEFAULT_PROTOCOLS' were commented out due to wrong token pricing. The update aims to ensure correct data is used for token exchange operations.