Refactor rpcProvider initialization in `getSDaiOracleTicker`
Replaced the direct use of `getRpcProvider` method with `getRpcNode` and `JsonRpcProvider` for creating the `rpcProvider` in `getSDaiOracleTicker` method. This change allows better error handling by throwing an error when RpcNode is not available, thus improving reliability and making sure the system provides helpful feedback when configuration or network issues arise.