Remove Docker scripts for project simplification
The numerous docker-related scripts and compose files are no longer useful in this project as they no longer follow the current configuration and deployment practices. These scripts are causing confusion and clutter hence they're deleted. The "docker-compose.yaml" was updated to reflect the removal of the multiply-proxy-actions service and some cleanup. In this way, the project is more straightforward and cleaner, thus improving the developer experience.