Remove deprecated Aave protocol data fetching
Deleted the logic for fetching Aave protocol data as it was deprecated and no longer in use. This includes deletion from the lending protocols, manage and open state machines, and ProductContext. Also, adjusted ViewPositionSectionComponent and ManageSectionComponent in the Aave to use currentPosition instead of protocolData for position information. Implemented these changes to clean up the codebase and improve efficiency.