Refactored TopBanners component and fixed typo
The unnecessary fragment wrapping the Box component in OneBanner function in the TopBanners component has been removed to simplify the component. This has no effect on the functionality or the appearance of the component. In addition, there was a typo in a variable name "loccationBasedBanner" which has been corrected to "locationBasedBanner". This makes the code cleaner and easier to read.