Client Country handling (#3013)
* Update handling of location-based banners and response headers

Modified middleware.ts and components/TopBanner.tsx to handle location-based banners. Changes made to use cookies instead of response headers for CloudFront-Viewer-Country. Renamed TopBanner.tsx to TopBanners.tsx and refactored the code to cater for multiple banners based on locations. Additionally, updated dependencies in yarn.lock.

The change was required to make the application more dynamic and personalized based on viewers' locations. With the use of cookies, user preferences can persist across sessions, providing a better user experience.

* Switched usePrevious hook from '@react-hooks-library/core' to 'react-use'

Changed the usePrevious hook import from '@react-hooks-library/core' to 'react-use' for wallet state components and safety reload features. The '@react-hooks-library/core' package has also been removed from the package.json file along with its dependencies from the yarn.lock file.size.

* Refactored TopBanners component and fixed typo

The unnecessary fragment wrapping the Box component in OneBanner function in the TopBanners component has been removed to simplify the component. This has no effect on the functionality or the appearance of the component. In addition, there was a typo in a variable name "loccationBasedBanner" which has been corrected to "locationBasedBanner". This makes the code cleaner and easier to read.