Add support for optimization tab on Aave V3 strategies.
This commit introduces the "isOptimizationTabEnabled" method to the Aave V3 strategy. This method checks if the optimization tab is enabled for the specific strategy type and network. This check is needed to separate the availability of the optimization tab based on the vault, as the feature is not yet supported on all vaults.

Further, it adds the feature toggles for "AUTO_BUY" and "AUTO_SELL" features under the Aave protocol in the automation feature map.