Add new lambdas and local development guide
This commit introduces two new lambda functions, 'get-triggers' and 'setup-trigger', and updates the documentation to include a guide for local development. Modifications were made in 'build.mjs' and 'install.mjs' files to accommodate the new lambdas, and necessary testing configurations and package-lock.json files were added. A significant part of the change went into updating the README with new instructions for creating a new lambda and to detail instructions for local development using AWS SAM and Docker.