Refactor portfolio domain-types import paths
Portfolio domain-types import paths have been updated for various components and helpers. A domain-types file has been created in the components/portfolio/types directory and the updated import paths reflect this change. Also, it corrects a pair of property names in aave features, and removes an unused variable declaration in AutoSellSidebarAaveVault and AutoBuySidebarAaveVault.