Update currency denomination in sidebars and details view
The update involves using a newly created function, getDenomination, from the calculations service. This function is used to set the currency code in both the AutoSellSidebarAaveVault and AutoBuySidebarAaveVault sidebars. It's also introduced in BasicAutomationDetailsView to show the correct denomination in the views. A minor adjustment in OptimizationControl's nextPrice calculation is dually covered.