Add custom RPC support for migrations API calls
The migrations API calls have been updated in the getAssetsForMigration helper and migrationsClient to support custom RPC URLs. Now, if the network is a custom fork, the network's RPC URL is passed as a parameter to the getPortfolioMigrationsResponse function. This allows using a custom network's RPC for the migration feature.